Update Room.java
Modify the Room class to add checkIn(), checkout() and cleanroom() methods to the room.

Once a room has been checked in, it should be occupied and marked as dirty. When a guest checks out of a room it must first be cleaned by a housekeeper before another guest can check into the room.